Writing rows to outfile w/ padding

+++ image.h
@@ -1,14 +1,16 @@
 #include <stdint.h>
+#include <sys/types.h>
 
 #ifndef IMAGE
-	#define IMAGE 1
+#define IMAGE 1
 #endif
 
 #ifndef PIXEL
-	#include "pixel.h"
+#include "pixel.h"
 #endif
 
-typedef struct __attribute__((__packed__)) BmpFileHeader {
+typedef struct __attribute__((__packed__)) BmpFileHeader
+{
     union
     {
         char signature[2];
@@ -20,7 +22,8 @@ typedef struct __attribute__((__packed__)) BmpFileHeader {
     uint32_t dataOffset;
 } BmpFileHeader;
 
-typedef struct __attribute__((__packed__)) DibHeader {
+typedef struct __attribute__((__packed__)) DibHeader
+{
     uint32_t headerSize;
     int32_t width;
     int32_t height;
@@ -28,32 +31,32 @@ typedef struct __attribute__((__packed__)) DibHeader {
     uint16_t bitsPerPixel;
     uint32_t compressionMethod;
     uint32_t rawImageSize;
-    int32_t  horRes;
-    int32_t  verRes;
+    int32_t horRes;
+    int32_t verRes;
     uint32_t numColors;
     uint32_t numImportantColors;
 } DibHeader;
 
-typedef struct bmp_Pixel24_t {
-	BmpFileHeader *bmpHeader;
-	DibHeader *dibHeader;
-	uint32_t postHeaderDataSize;
-	char *postHeaderData;
-	Pixel24_t *pixelArray;
-	uint32_t postDataSize;
-	char *postData;
+typedef struct bmp_Pixel24_t
+{
+    BmpFileHeader *bmpHeader;
+    DibHeader *dibHeader;
+    uint32_t postHeaderDataSize;
+    char *postHeaderData;
+    Pixel24_t *pixelArray;
+    uint32_t postDataSize;
+    char *postData;
 } bmp_Pixel24_t;
 
-
 // Pixel24_t type functions
-size_t readBitmapPixelData(bmp_Pixel24_t *dest, FILE* bitmap);
-size_t removeRedPixel24_t(bmp_Pixel24_t *bitmap);
-size_t removeBluePixel24_t(bmp_Pixel24_t *bitmap);
-size_t removeGreenPixel24_t(bmp_Pixel24_t *bitmap);
-size_t randomChannelPixel24_t(bmp_Pixel24_t *bitmap, ColorChannel channel);
+ssize_t readBitmapPixelData(bmp_Pixel24_t *dest, FILE *bitmap);
+ssize_t removeRedPixel24_t(bmp_Pixel24_t *bitmap);
+ssize_t removeBluePixel24_t(bmp_Pixel24_t *bitmap);
+ssize_t removeGreenPixel24_t(bmp_Pixel24_t *bitmap);
+ssize_t randomChannelPixel24_t(bmp_Pixel24_t *bitmap, ColorChannel channel);
 
 // Assumes pixelArray field has already been allocated
-unsigned readAllRows_Pixel24_t(bmp_Pixel24_t *dest, FILE* bitmap);
+ssize_t readAllRows_Pixel24_t(bmp_Pixel24_t *dest, FILE *bitmap);
 
-size_t writeToFilePixel24_t(bmp_Pixel24_t *bitmap, FILE* outFile);
+ssize_t writeToFilePixel24_t(bmp_Pixel24_t *bitmap, FILE *outFile);
 int freeImageDataPixel24_t(bmp_Pixel24_t *bitmap);
